pyproject.toml
[tool.poetry]
name = "projectr"
version = "0.1.0"
description = ""
authors = ["Lorenzo Balugani <lorenzo.balugani@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.5"
Django = "^3.2"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
